DBLoader: clean-up and delete & create train table on each run
